# swagger.yaml
swagger: "2.0"
info:
  description: "TVJ Internal tool"
  version: "1.0.0"
  title: "TVJ Internal tool"
host: "localhost"
...
schemes:
- "https"
- "http"
paths:
...
  /member-management/search:
    get:
      tags:
      - "Member Management"
      summary: "Search member"
      description: "Search member"
      operationId: "searchMember"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "search by name, between 0 and 40 characters"
        type: "string"
      - name: "username"
        in: "query"
        description: "search by username, between 0 and 20 characters"
        type: "string"
      - name: "titleId"
        in: "query"
        description: "search by title, between 0 and 40 characters"
        type: "string"
      - name: "isActivated"
        in: "query"
        description: "search by is activated"
        type: "boolean"
      - name: "limit"
        in: "query"
        description: "paging limit, min = 0"
        type: "integer"
      - name: "offset"
        in: "query"
        description: "paging offset, min = 1"
        type: "integer"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/MemberListResDto"
  /member-management/add-member:
    post:
      tags:
      - "Member Management"
      summary: "Add member"
      description: "Add member"
      operationId: "addMember"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "MemberAddReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/MemberAddReqDto"
      responses:
        200:
          description: "Error code: SUC006"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Error code: ERR012"
          schema:
            $ref: "#/definitions/MessageResDto"
  /member-management/update-member:
    put:
      tags:
      - "Member Management"
      summary: "Update member"
      description: "Update member"
      operationId: "updateMember"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "MemberUpdateReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/MemberUpdateReqDto"
      responses:
        200:
          description: "Error code: SUC007"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Error code: ERR013"
          schema:
            $ref: "#/definitions/MessageResDto"
  /member-management/view-member:
    get:
      tags:
      - "Member Management"
      summary: "View member"
      description: "View member"
      operationId: "viewMember"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "search by username, less than 20 characters"
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/MemberResDto"
        400:
          description: "Error code: ERR014"
          schema:
            $ref: "#/definitions/MessageResDto"
  /member-management/update-member-activate-status:
    patch:
      tags:
      - "Member Management"
      summary: "Update member activate status"
      description: "Update member activate status"
      operationId: "updateMemberActivateStatus"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "MemberActivateStatusUpdateReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/MemberActivateStatusUpdateReqDto"
      responses:
        200:
          description: "Error code: SUC008"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Error code: ERR015"
          schema:
            $ref: "#/definitions/MessageResDto"
  /member-management/delete-member:
    delete:
      tags:
      - "Member Management"
      summary: "Delete member"
      description: "Delete member"
      operationId: "deleteMember"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "body"
          name: "MemberDeleteReqDto"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/MemberDeleteReqDto"
      responses:
        200:
          description: "Error code: SUC009"
          schema:
            $ref: "#/definitions/MessageResDto"
        400:
          description: "Error code: ERR016"
          schema:
            $ref: "#/definitions/MessageResDto"
definitions:
...
  MemberListResDto:
    type: "object"
    properties:
      memberResDtoList:
        type: "array"
        items:
          $ref: "#/definitions/MemberResDto"
  MemberResDto:
    type: "object"
    properties:
      username:
        type: string
        description: ""    
      firstName:
        type: string
        description: ""   
      lastName:
        type: string
        description: ""    
      email:
        type: string
        description: ""   
      titleId:
        type: string
        description: ""    
      titleName:
        type: string
        description: ""    
      activated:
        type: boolean
        description: ""       
  MemberAddReqDto:
    type: "object"
    properties:
      username:
        type: string
        description: "Not Blank, less than 20 characters"
      firstName:
        type: string
        description: "Not Blank"
      lastName:
        type: string
        description: "Not Blank"
      titleId:
        type: string
        description: "Not Blank, less than 40 characters"
      email:
        type: string
        description: "Not Blank, must have domain tinhvan.com"
      password:
        type: string
        description: "Not Blank"
  MemberUpdateReqDto:
    type: "object"
    properties:
      username:
        type: string
        description: "Not Blank, less than 20 characters"
      firstName:
        type: string
        description: "Not Blank"
      lastName:
        type: string
        description: "Not Blank"
      titleId:
        type: string
        description: "Not Blank, less than 40 characters"
      email:
        type: string
        description: "Not Blank, must have domain tinhvan.com"
  MemberActivateStatusUpdateReqDto:      
    type: "object"
    properties:
      username:
        type: string
        description: "Not Blank, less than 20 characters"
      activated:
        type: boolean
        description: "Not Blank"    
  MemberDeleteReqDto:      
    type: "object"
    properties:
      username:
        type: string
        description: "Not Blank, less than 20 characters"
  MessageResDto:
    type: "object"
    properties:
      code:
        type: string
        description: ""
      message:
        type: string
        description: ""